Data Secure CemiFrame handler
